services/loader_deployed/pyproject.toml
[project]
name = "minimal-rag-ingest"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "feedparser>=6.0.10",
  "httpx>=0.27.0",
  "readability-lxml>=0.8.1",
  "lxml>=5.2.1",
  "psycopg[binary]>=3.2.1",
  "python-dateutil>=2.9.0",
  "langdetect>=1.0.9",
  "trafilatura>=1.9.0",
  "sentence-transformers>=5.1.1",
  "pgvector>=0.4.1",
  "pandas>=2.3.3",
  "google-genai>=1.43.0",
  "google-cloud-storage>=3.4.1",
  "langchain>=0.3.27",
  "langchain-experimental>=0.3.4",
  "langchain-openai>=0.3.35",
  "langchain-huggingface>=0.3.1",
  "numpy>=2.3.3",
  "langchain-text-splitters>=0.0.1",
  "fastapi>=0.104.1",
  "uvicorn[standard]>=0.24.0",
  "dvc>=3.63.0",
  "dvc-gs>=3.0.2",
]
